a little documentation
